# src/app/api/mcp/maze-game/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "MCPlayerOne Game API",
    "version": "1.0.0",
    "description": "API for interacting with the MCP text adventure game. Allows starting games, getting state, moving, taking items, examining things, and viewing leaderboards."
  },
  "servers": [
    {
      "url": "/api/game",
      "description": "Relative path to the game API endpoints"
    }
  ],
  "paths": {
    "/start": {
      "post": {
        "summary": "Start Game / Get Initial State",
        "description": "Starts a new game session for a user in a specific story or retrieves the existing state if the user has already started.",
        "operationId": "startGame",
        "tags": ["Game Actions"],
        "requestBody": {
          "description": "User and Story identifiers.",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserStoryInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Game started or state retrieved successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameStateResponse"
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/state": {
      "post": {
        "summary": "Get Current Game State",
        "description": "Retrieves the current state of the game (player and location) for a specific user and story. Requires `userId` and `storyId`. Obtain `storyId` using the `listStories` tool. Obtain `userId` from the result of `startGame` (for new players) or user input (for resuming). NOTE: The current MCP tool definition for `getGameState` may be faulty and might not accept the required `userId` and `storyId` parameters, potentially only accepting `random_string`. A correctly defined tool *must* send `userId` and `storyId`.",
        "operationId": "getGameState",
        "tags": ["Game Actions"],
        "requestBody": {
          "description": "User and Story identifiers.",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserStoryInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Current game state retrieved successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameStateResponse"
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
...
    "/stories/{storyId}": {
      "get": {
        "summary": "Get Story By ID",
        "description": "Retrieves the details of a specific story using its logical ID.",
        "operationId": "getStoryById",
        "tags": ["Game Setup"],
        "parameters": [
          {
            "name": "storyId",
            "in": "path",
            "required": true,
            "description": "The unique logical ID of the story to retrieve.",
            "schema": {
              "type": "string"
            },
             "example": "mystic_library"
          }
        ],
        "responses": {
          "200": {
            "description": "Story details retrieved successfully.",
            "content": {
              "application/json": {
                 "schema": { "$ref": "#/components/schemas/StoryRecord" }
              }
            }
          },
           "400": { "$ref": "#/components/responses/BadRequest" },
           "404": { "$ref": "#/components/responses/NotFound" },
           "500": { "$ref": "#/components/responses/InternalError" }
        }
      },
      "delete": {
        "summary": "Delete Story",
        "description": "Deletes a specific story and all associated data (locations, items, player states) from the database.",
        "operationId": "deleteStory",
        "tags": ["Game Setup"],
        "parameters": [
          {
            "name": "storyId",
            "in": "path",
            "required": true,
            "description": "The unique ID of the story to delete.",
            "schema": {
              "type": "string"
            },
             "example": "my_generated_story"
          }
        ],
        "responses": {
          "200": {
            "description": "Story and associated data deleted successfully.",
            "content": {
              "application/json": {
                "schema": {
                   "type": "object",
                   "properties": {
                     "success": { "type": "boolean", "example": true },
                     "message": { "type": "string", "example": "Story 'my_generated_story' and all associated data deleted successfully." }
                   }
                 }
              }
            }
          },
           "400": { "$ref": "#/components/responses/BadRequest" },
           "404": { "$ref": "#/components/responses/NotFound" },
           "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "UserStoryInput": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "description": "The unique identifier for the player.",
            "example": "user123"
          },
          "storyId": {
            "type": "string",
            "description": "The unique identifier for the story.",
            "example": "dragon_lair"
          }
        },
        "required": ["userId", "storyId"]
      },
...
      "PlayerState": {
        "type": "object",
        "description": "Represents the state of a player within a story.",
        "properties": {
          "_id": { "type": "string", "description": "Internal document ID (storyId_userId).", "example": "dragon_lair_user123" },
          "userId": { "type": "string", "example": "user123" },
          "storyId": { "type": "string", "example": "dragon_lair" },
          "currentLocation": { "type": "string", "description": "ID of the player's current location.", "example": "dragon_entrance" },
          "inventory": { 
            "type": "array", 
            "items": { "type": "string" }, 
            "description": "List of item IDs in the player's inventory.",
            "example": ["rusty_sword"]
          },
           "discoveredLocations": { 
            "type": "array", 
            "items": { "type": "string" }, 
            "description": "List of location IDs the player has visited.",
            "example": ["dragon_entrance", "guard_chamber"]
          },
          "gameProgress": {
            "type": "object",
            "properties": {
              "itemsFound": { "type": "array", "items": { "type": "string" }, "example": ["rusty_sword", "dragon_scale"]},
              "puzzlesSolved": { "type": "array", "items": { "type": "string" }, "example": ["gate_puzzle"]},
              "storyProgress": { "type": "number", "format": "int32", "example": 50 }
            }
          }
        }
      },
      "Location": {
        "type": "object",
        "description": "Represents a location within a story. If an image is present, both a top-level 'image' and 'alt' field and a 'content' array are included. The content array contains a type: 'image' object first, followed by a type: 'text' object.",
        "properties": {
          "id": { "type": "string", "description": "Unique ID for the location within the story.", "example": "dragon_entrance" },
          "storyId": { "type": "string", "example": "dragon_lair" },
          "name": { "type": "string", "example": "Dragon Entrance" },
          "description": { "type": "string", "example": "You stand before a dark cave entrance..." },
          "items": { 
            "type": "array", 
            "items": { "type": "string" }, 
            "description": "List of item IDs currently present in the location.",
            "example": ["rusty_sword"] 
          },
           "features": { 
             "type": "array", 
             "items": { 
               "type": "object", 
               "properties": {
                 "id": {"type": "string", "example": "large_door"},
                 "name": {"type": "string", "example": "Large Door"},
                 "description": {"type": "string", "example": "A heavy oak door..."}
               }
             }, 
             "description": "List of examinable features in the location."
          },
          "exits": { 
            "type": "array", 
            "items": { "type": "string" }, 
            "description": "List of location IDs accessible from this location.",
            "example": ["guard_chamber"]
          },
           "requirements": {
             "type": "object",
             "properties": {
               "item": { "type": "string", "nullable": true, "description": "Item ID required to enter.", "example": "silver_key" },
               "condition": { "type": "string", "nullable": true, "description": "Condition ID (e.g., puzzle) required to enter.", "example": "gate_puzzle" }
             }
           },
          "image": { "type": "string", "description": "URL of the image for the location, if present.", "nullable": true },
          "alt": { "type": "string", "description": "Alt text for the image, if present.", "nullable": true },
          "content": {
            "type": "array",
            "items": {
              "oneOf": [
                { "$ref": "#/components/schemas/ImageContent" },
                { "$ref": "#/components/schemas/TextContent" }
              ]
            },
            "description": "Array with image and text content for the location."
          }
        }
      },
      "GameStateResponse": {
        "type": "object",
        "properties": {
           "success": { "type": "boolean", "example": true },
           "storyId": { "type": "string", "description": "The unique identifier for the story." },
           "userId": { "type": "string", "description": "The unique identifier for the user/player." },
           "message": { "type": "string", "nullable": true, "example": "Welcome back!" },
           "player": { "$ref": "#/components/schemas/PlayerState" },
           "location": { "$ref": "#/components/schemas/Location" }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "string",
            "description": "A message describing the error."
          },
          "hint": {
             "type": "string",
             "nullable": true,
             "description": "An optional hint for the user/agent."
          }
        },
        "required": ["success", "error"]
      },
      "StoryRecord": {
        "properties": {
          "_id": { "type": "string", "description": "Internal database ID.", "readOnly": true },
          "storyId": { "type": "string", "description": "Unique logical story identifier (echoed at top level for agent workflows).", "example": "mystic_library" },
          "title": { "type": "string", "example": "The Mystic Library", "description": "Short, punchy, creative title. No prefixes or bylines." },
          "description": { "type": "string", "example": "An ancient library filled with magical secrets." },
          "startingLocation": { "type": "string", "example": "entrance_hall" },
          "version": { "type": "string", "example": "1.1" },
          "requiredArtifacts": {
            "type": "array",
            "items": { "type": "string" },
            "description": "List of item IDs that must be collected to win. No more than the value of MAX_REQUIRED_ARTIFACTS (from env, default 5) are required; if more are generated, a random subset will be chosen."
          },
          "goalRoomId": {
            "type": "string",
            "description": "The ID of the goal room. To win, a player must collect all requiredArtifacts and reach this room. Backend validates that a win path exists after world generation."
          },
          "theme": { "type": "string", "description": "Theme used for generation (optional)." },
          "image": { "type": "string", "description": "URL of the story cover image, if present.", "nullable": true },
          "alt": { "type": "string", "description": "Alt text for the story image, if present.", "nullable": true },
          "content": {
            "type": "array",
            "items": {
              "oneOf": [
                { "$ref": "#/components/schemas/ImageContent" },
                { "$ref": "#/components/schemas/TextContent" }
              ]
            },
            "description": "Array with image and text content for the story."
          }
        },
        "required": ["id", "title", "description", "startingLocation", "version"]
      },
...
      "ImageContent": {
        "type": "object",
        "properties": {
          "type": { "type": "string", "enum": ["image"] },
          "image": { "type": "string", "description": "URL of the image." },
          "alt": { "type": "string", "description": "Alt text for the image." }
        }
      },
      "TextContent": {
        "type": "object",
        "properties": {
          "type": { "type": "string", "enum": ["text"] },
          "text": { "type": "string", "description": "JSON stringified object with all metadata, including storyId." }
        }
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "Bad Request - The request was malformed or missing required parameters.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
...
      "NotFound": {
        "description": "Not Found - The requested resource (player, story, location, item) could not be found.",
         "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
      "InternalError": {
        "description": "Internal Server Error - Something went wrong on the server.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      }
    }
  }
} 